+++ lib/inertia_rails/helper.rb
@@ -23,9 +23,10 @@ module InertiaRails
 
     def inertia_meta_tags
       meta_tag_data = (inertia_page || {}).dig(:props, :_inertia_meta) || []
+      head_attribute = controller.send(:inertia_configuration).head_attribute
 
       meta_tags = meta_tag_data.map do |inertia_meta_tag|
-        inertia_meta_tag.to_tag(tag)
+        inertia_meta_tag.to_tag(tag, inertia_attribute: head_attribute)
       end
 
       safe_join(meta_tags, "\n")